transfer money

+++ src/db/card.js
@@ -1,25 +1,27 @@
 import { execSql } from "./utils";
 import { stripObj } from "../services";
-import { createTransaction } from "./transaction";
+import { createTransaction, updateInitTransactionCash } from "./transaction";
 
 export async function createCard({ name, type, money, note }) {
-  const [
-    ,
-    { insertId },
-  ] = await execSql(`insert into cards (name, type, note) values (?,?,?)`, [
-    name,
-    type,
-    note,
-  ]);
-  await createTransaction({
-    category: 1,
-    card: insertId,
-    cash: money,
-    date: new Date().toLocaleDateString(),
-    note: `Init card ${insertId}`,
-  });
-  console.log("finish create Card");
-  return await getCard();
+  try {
+    const [
+      ,
+      { insertId },
+    ] = await execSql(`insert into cards (name, type, note) values (?,?,?)`, [
+      name,
+      type,
+      note,
+    ]);
+    return await createTransaction({
+      category: 1,
+      card: insertId,
+      cash: money,
+      date: new Date().toLocaleDateString(),
+      note: `Init card ${insertId}`,
+    });
+  } catch (err) {
+    console.log(err);
+  }
 }
 
 export async function getCard() {
@@ -40,21 +42,73 @@ export async function getCardById(id) {
   }
 }
 
-export async function updateCard(data) {
-  if (data.id) {
-    const oldData = await getCardById(data.id);
-    const { name, type, note, id } = { ...oldData, ...stripObj(data) };
-    await execSql(`update cards set name=?, type=?, note=?  where id=?`, [
-      name,
-      type,
-      note,
-      id,
-    ]);
+export async function countCard() {
+  try {
+    const [, { rows }] = await execSql(
+      `select count(*) as count from cards`,
+      []
+    );
+    return rows.item(0).count;
+  } catch (err) {
+    console.log(err);
   }
-  return await getCard();
+}
+
+export async function updateCard(data) {
+  if (data.id)
+    try {
+      const beforeUpdated = await getCardById(data.id);
+      const { name, type, note, id, money } = {
+        ...beforeUpdated,
+        ...stripObj(data),
+      };
+      await execSql(`update cards set name=?, type=?, note=?  where id=?`, [
+        name,
+        type,
+        note,
+        id,
+      ]);
+      return await updateInitTransactionCash({
+        card: id,
+        cash: money,
+      });
+    } catch (err) {
+      console.log(err);
+    }
 }
 
 export async function deleteCard(id) {
-  await execSql(`delete from cards where id=?  `, [id]);
-  return await getCard();
+  if ((await countCard()) > 1)
+    try {
+      const beforeDeleted = await getCardById(id);
+      await execSql(`delete from cards where id=?  `, [id]);
+      return beforeDeleted;
+    } catch (err) {
+      console.log(err);
+    }
 }
+
+export const transferMoney = async ({ sendId, receiveId, money, note }) => {
+  try {
+    const sender = await createTransaction({
+      category: 2,
+      card: sendId,
+      cash: -money,
+      date: new Date().toLocaleDateString(),
+      note: `(Transfer money to card ${receiveId}) ${note}`,
+    });
+    const recipient = await createTransaction({
+      category: 2,
+      card: receiveId,
+      cash: money,
+      date: new Date().toLocaleDateString(),
+      note: `(Receive money from card ${sendId}) ${note}`,
+    });
+    return {
+      sender,
+      recipient,
+    };
+  } catch (err) {
+    console.log(err);
+  }
+};
